Fixed a spurious bug caused by offset calculation
`ArrayAccess` previously used the class type's size for determining array offsets, which is wrong.
